+++ lib/luadata.h
@@ -17,7 +17,7 @@ LUNATIK_LIB(data);
 
 #define luadata_clear(o)	(luadata_reset((o), NULL, 0, LUADATA_OPT_KEEP))
 
-lunatik_object_t *luadata_new(void *ptr, size_t size, bool sleep, uint32_t opt);
+lunatik_object_t *luadata_new(void *ptr, size_t size, bool sleep, uint8_t opt);
 int luadata_reset(lunatik_object_t *object, void *ptr, size_t size, uint8_t opt);
 
 static inline void luadata_close(lunatik_object_t *object)
